requirements.txt
utils~=1.0.2
loguru~=0.7.3
requests~=2.32.5
python-telegram-bot~=22.5